Minor user service bug fixes

+++ user-service/src/main/java/com/iroit/user_service/services/UserService.java
@@ -1,88 +1,72 @@
 package com.iroit.user_service.services;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.iroit.user_service.exceptions.InvalidRequestException;
+import com.iroit.user_service.exceptions.ResourceNotFoundException;
 import com.iroit.user_service.models.User;
 import com.iroit.user_service.repositories.UserRepository;
 
 @Service
 public class UserService {
-  
+
   @Autowired
   private UserRepository userRepository;
 
 //#region Get methods
-  public List<User> getUsers() throws Exception {
-    List<User> result = userRepository.findAll();
-
-    if (result == null || result.size() == 0)
-      throw new Exception("No users found!");
-
-    return result;
+  public List<User> getUsers() {
+    return userRepository.findAll();
   }
 
-  public User getUserById(Long userId) throws Exception {
+  public User getUserById(Long userId) {
     if (userId == null || userId < 0)
-      throw new Exception("Invalid ID supplied!");
+      throw new InvalidRequestException("Invalid ID supplied!");
 
-    Optional<User> result = userRepository.findById(userId);
-
-    if (result.isEmpty())
-      throw new Exception("The user was not found!");
-
-    return result.get();
+    return userRepository.findById(userId)
+        .orElseThrow(() -> new ResourceNotFoundException("The user was not found!"));
   }
 //#endregion
 
 //#region Post methods
-  public User createUser(User user) throws Exception {
-
+  public User createUser(User user) {
     if (user == null)
-      throw new Exception("Invalid user data supplied", new Throwable("Data"));
+      throw new InvalidRequestException("Invalid user data supplied!");
 
-    User result = userRepository.save(user);
-
-    if (result == null)
-      throw new Exception("Couldn't create the user!");
-
-    return result;
+    return userRepository.save(user);
   }
 //#endregion
 
 //#region Put methods
-  public User updateUser(Long userId, User user) throws Exception {
+  public User updateUser(Long userId, User user) {
     if (userId == null || userId < 0)
-      throw new Exception("Invalid ID supplied!", new Throwable("ID"));
+      throw new InvalidRequestException("Invalid ID supplied!");
 
     if (user == null)
-      throw new Exception("Invalid user data supplied", new Throwable("Data"));
+      throw new InvalidRequestException("Invalid user data supplied!");
 
-    User result = userRepository.findById(userId).map(foundUser -> {
-      foundUser.setFirstName(user.getFirstName());
-      foundUser.setLastName(user.getLastName());
-      foundUser.setDateOfBirth(user.getDateOfBirth());
-      return userRepository.save(user);
-    }).orElseThrow(() -> new Exception("Couldn't update the user!"));
+    User foundUser = userRepository.findById(userId)
+        .orElseThrow(() -> new ResourceNotFoundException("The user was not found!"));
 
-    return result;
+    foundUser.setFirstName(user.getFirstName());
+    foundUser.setLastName(user.getLastName());
+    foundUser.setDateOfBirth(user.getDateOfBirth());
+
+    return userRepository.save(foundUser);
   }
 //#endregion
 
 //#region Delete methods
-  public void deleteUser(Long userId) throws Exception {
+  public void deleteUser(Long userId) {
     if (userId == null || userId < 0)
-      throw new Exception("Invalid ID supplied!", new Throwable("ID"));
+      throw new InvalidRequestException("Invalid ID supplied!");
 
-    try {
-      userRepository.deleteById(userId);
-    } catch (Exception e) {
-      throw new Exception("Couldn't delete the user!");
-      // TODO add logging
-    }
+    if (!userRepository.existsById(userId))
+      throw new ResourceNotFoundException("The user was not found!");
+
+    userRepository.deleteById(userId);
   }
 //#endregion
